Fix some strcpy warnings

+++ hotspot/logic/requests/process_request.h
@@ -1,2 +1,2 @@
 
-int process_request(void *configuration, const char *request, const char *params, char *result, size_t max_result_len);
+size_t process_request(void *configuration, const char *request, const char *params, char *result, size_t max_result_len);
